+++ HISTORY.md
@@ -42,6 +42,7 @@
 * extensions: new extension primitives for hiding and showing arbitrary blocks in the palette
 * threads: keep hidden variables out of the palette and drop-down menus
 * objects: added utilities to enumerate all palette blocks for hiding & showing
+* objects, threads, extensions: refactored block hiding methods
 
 ### 2021-10-05
 * threads, store: added infrastructure for hiding individual variables in palette